# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * HVCG Lead Magnet Builder API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type LeadMagnetType = typeof LeadMagnetType[keyof typeof LeadMagnetType];


export const LeadMagnetType = {
  existing_url: 'existing_url',
  give_away: 'give_away',
} as const;

export type LeadMagnetStatus = typeof LeadMagnetStatus[keyof typeof LeadMagnetStatus];


export const LeadMagnetStatus = {
  draft: 'draft',
  review: 'review',
  live: 'live',
} as const;

export interface LeadMagnet {
  id: number;
  userId: string;
  type: LeadMagnetType;
  status: LeadMagnetStatus;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  existingUrl?: string | null;
  /** @nullable */
  businessName?: string | null;
  /** @nullable */
  businessLocation?: string | null;
  /** @nullable */
  giveawayFileName?: string | null;
  /** @nullable */
  giveawayFileUrl?: string | null;
  /** @nullable */
  templateId?: number | null;
  /** @nullable */
  customFontColor?: string | null;
  /** @nullable */
  customBgColor?: string | null;
  /** @nullable */
  customTextColor?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  tagline?: string | null;
  /** @nullable */
  shareUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type LeadMagnetInputType = typeof LeadMagnetInputType[keyof typeof LeadMagnetInputType];


export const LeadMagnetInputType = {
  existing_url: 'existing_url',
  give_away: 'give_away',
} as const;

export interface LeadMagnetInput {
  type: LeadMagnetInputType;
  title?: string;
  description?: string;
  existingUrl?: string;
  businessName?: string;
  businessLocation?: string;
  templateId?: number;
  customFontColor?: string;
  customBgColor?: string;
  customTextColor?: string;
}

export interface LeadMagnetUpdate {
  title?: string;
  description?: string;
  existingUrl?: string;
  businessName?: string;
  businessLocation?: string;
  templateId?: number;
  customFontColor?: string;
  customBgColor?: string;
  customTextColor?: string;
  logoUrl?: string;
  tagline?: string;
}

export interface LeadMagnetsSummary {
  total: number;
  live: number;
  review: number;
  draft: number;
}

export type TemplateLayout = typeof TemplateLayout[keyof typeof TemplateLayout];


export const TemplateLayout = {
  split: 'split',
  banner: 'banner',
  overlay: 'overlay',
  stacked: 'stacked',
  minimal: 'minimal',
} as const;

export interface Template {
  id: number;
  name: string;
  layout: TemplateLayout;
  description: string;
  previewColor: string;
  accentColor: string;
  fontFamily: string;
  /** @nullable */
  sampleImageUrl?: string | null;
}

export interface Example {
  id: number;
  industry: string;
  title: string;
  description: string;
  giveawayType: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface Industry {
  id: number;
  name: string;
  slug: string;
}

export interface FileUploadInput {
  fileName: string;
  /** Base64-encoded data URL of the file */
  fileDataUrl: string;
}

export interface FileUploadResult {
  fileUrl: string;
  fileName: string;
}

export interface AiPrefillInput {
  /** @nullable */
  sourceUrl?: string | null;
  /**
     * Base64-encoded document content
     * @nullable
     */
  documentDataUrl?: string | null;
}

export interface AiPrefillResult {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  businessName?: string | null;
  /** @nullable */
  businessLocation?: string | null;
}

export interface AiBrandingInput {
  /** @nullable */
  sourceUrl?: string | null;
  /** @nullable */
  documentDataUrl?: string | null;
}

export interface AiBrandingResult {
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  tagline?: string | null;
  /** @nullable */
  primaryColor?: string | null;
}

export type ListExamplesParams = {
/**
 * @nullable
 */
industry?: string | null;
};