# lib/api-zod/src/generated/types/aiBrandingResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * HVCG Lead Magnet Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface AiBrandingResult {
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  tagline?: string | null;
  /** @nullable */
  primaryColor?: string | null;
}