# lib/api-zod/src/generated/types/aiPrefillInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * HVCG Lead Magnet Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface AiPrefillInput {
  /** @nullable */
  sourceUrl?: string | null;
  /**
     * Base64-encoded document content
     * @nullable
     */
  documentDataUrl?: string | null;
}