# lib/api-zod/src/generated/types/fileUploadInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * HVCG Lead Magnet Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface FileUploadInput {
  fileName: string;
  /** Base64-encoded data URL of the file */
  fileDataUrl: string;
}