# lib/api-zod/src/generated/types/leadMagnetInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * HVCG Lead Magnet Builder API
 * OpenAPI spec version: 0.1.0
 */
import type { LeadMagnetInputType } from './leadMagnetInputType';

export interface LeadMagnetInput {
  type: LeadMagnetInputType;
  title?: string;
  description?: string;
  existingUrl?: string;
  businessName?: string;
  businessLocation?: string;
  templateId?: number;
  customFontColor?: string;
  customBgColor?: string;
  customTextColor?: string;
}