# lib/api-zod/src/generated/types/leadMagnetsSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * HVCG Lead Magnet Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface LeadMagnetsSummary {
  total: number;
  live: number;
  review: number;
  draft: number;
}