# lib/api-zod/src/generated/types/template.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * HVCG Lead Magnet Builder API
 * OpenAPI spec version: 0.1.0
 */
import type { TemplateLayout } from './templateLayout';

export interface Template {
  id: number;
  name: string;
  layout: TemplateLayout;
  description: string;
  previewColor: string;
  accentColor: string;
  fontFamily: string;
  /** @nullable */
  sampleImageUrl?: string | null;
}